Frame interpolation and dot lottie support (#3)
* Update the doc

* Update thorvg

* Add Interpolation param

* Set Interpolation param

* Set a custom measure

* Set Custom Mode

* Add color parser (from string to color)

* Set color background

* Update common

* Update ThorVG

* Update Sample app

* Add dot lottie support

* Add maven publish plugin

* Add lib config

* Add foojay for Java compat

* Remote java compat

* Update AGP

* Update AGP

* Update AGP

* Update Java

* Update Readme

* Update Readme

* Convert isPlaying(), isPaused() and isStopped() method into property

* Expose the backgroundColor Property

* Implement destroy method

* Expose the direction API and prevent from going our the frame

* Add support json string content as data

* Refactoring

* Load a dot lottie as byte array

+++ thorvg-android/src/main/java/com/lottiefiles/dotlottie/core/LottieNative.kt
@@ -8,11 +8,11 @@ object LottieNative {
     }
 
     @JvmStatic
-    external fun nCreateLottie(content: String?, length: Int, outValues: IntArray?): Long
+    external fun nCreateLottie(content: String?, length: Int, outValues: DoubleArray?): Long
     @JvmStatic
     external fun nSetLottieBufferSize(lottiePtr: Long, bitmap: Bitmap?, width: Float, height: Float)
     @JvmStatic
-    external fun nDrawLottieFrame(lottiePtr: Long, bitmap: Bitmap?, frame: Int)
+    external fun nDrawLottieFrame(lottiePtr: Long, bitmap: Bitmap?, frame: Float)
     @JvmStatic
     external fun nDestroyLottie(lottiePtr: Long)
 }
